split netcdf and other store
