+++ openpathsampling/analysis/move_scheme.py
@@ -3,7 +3,7 @@ import openpathsampling as paths
 from openpathsampling.analysis.move_strategy import levels as strategy_levels
 import openpathsampling.analysis.move_strategy as strategies
 
-from openpathsampling.base import StorableNamedObject
+from openpathsampling.netcdfplus.base import StorableNamedObject
 
 try:
     import pandas as pd
