+++ openpathsampling/chaindict.py
@@ -3,10 +3,9 @@ __author__ = 'Jan-Hendrik Prinz'
 import collections
 
 import numpy as np
-import weakref
 
-from openpathsampling.storage.cache import LRUCache
-from openpathsampling.storage.objproxy import LoaderProxy
+from openpathsampling.netcdfplus.cache import LRUCache
+from openpathsampling.netcdfplus.objproxy import LoaderProxy
 
 
 class ChainDict(object):
