+++ openpathsampling/dynamics_engine.py
@@ -10,7 +10,7 @@ import logging
 import simtk.unit as u
 
 import openpathsampling as paths
-from openpathsampling.base import StorableNamedObject
+from openpathsampling.netcdfplus.base import StorableNamedObject
 
 logger = logging.getLogger(__name__)
