+++ openpathsampling/netcdfplus/base.py
