+++ openpathsampling/pathmover.py
@@ -11,10 +11,9 @@ import logging
 import numpy as np
 
 import openpathsampling as paths
-from openpathsampling.base import StorableNamedObject, StorableObject
+from openpathsampling.netcdfplus.base import StorableNamedObject, StorableObject
 from ops_logging import initialization_logging
 from treelogic import TreeMixin
-
 
 logger = logging.getLogger(__name__)
 init_log = logging.getLogger('openpathsampling.initialization')
