+++ openpathsampling/snapshot.py
@@ -12,7 +12,7 @@ import mdtraj as md
 # THIS SHOULD NOT BE NECESSARY! SEE XYZ BELOW
 import simtk.unit as u
 
-from openpathsampling.base import StorableObject, lazy_loading_attributes
+from openpathsampling.netcdfplus.base import StorableObject, lazy_loading_attributes
 
 
 # =============================================================================
