+++ openpathsampling/storage/__init__.py
@@ -1,6 +1,4 @@
 from storage import Storage, AnalysisStorage
-from openpathsampling.storage.object_json import StorableObjectJSON
-from object_storage import ObjectStore
 from snapshot_store import SnapshotStore, ConfigurationStore, MomentumStore
 from trajectory_store import TrajectoryStore
 from sample_store import SampleStore, SampleSetStore
