+++ openpathsampling/storage/mcstep_store.py
@@ -1,4 +1,4 @@
-from openpathsampling.storage import ObjectStore
+from openpathsampling.netcdfplus.object_storage import ObjectStore
 from openpathsampling.pathsimulator import MCStep
