+++ openpathsampling/storage/sample_store.py
@@ -1,6 +1,6 @@
-from object_storage import ObjectStore
+from openpathsampling.netcdfplus.object_storage import ObjectStore
 from openpathsampling.sample import SampleSet, Sample
-from objproxy import LoaderProxy
+from openpathsampling.netcdfplus.objproxy import LoaderProxy
 
 
 class SampleStore(ObjectStore):
