+++ openpathsampling/tests/teststorage.py
@@ -2,23 +2,17 @@
 @author David W.H. Swenson
 '''
 import os
-from nose.tools import (assert_equal, assert_not_equal, assert_items_equal,
-                        assert_almost_equal, raises)
-from nose.plugins.skip import Skip, SkipTest
-from test_helpers import (true_func,
-                          data_filename,
-                          assert_equal_array_array,
-                          assert_not_equal_array_array,
+
+from nose.tools import (assert_equal)
+import mdtraj as md
+
+from test_helpers import (data_filename,
                           compare_snapshot
                           )
-
 from openpathsampling.openmm_engine import *
 from openpathsampling.snapshot import Snapshot
-from openpathsampling.storage.object_json import ObjectJSON
-
+from openpathsampling.netcdfplus.object_json import ObjectJSON
 from openpathsampling.storage import Storage
-
-import mdtraj as md
 
 
 class testStorage(object):
